the method widens 26 notorch-vlm-2m alloc/copy products to size_t before int overflow
CodeQL flagged 26 cpp/integer-multiplication-cast-to-long alerts across the vendored
vision stack, each an int*int product overflowing before the implicit widen to size_t.
Leading operand cast to (size_t):

  ariannamethod/notorch.c        20  tape-backward calloc: T*D, rows*cols, T*V, ...
  ariannamethod/notorch_vision.h  4  image CHW malloc: c*h*w, channels*th*tw
  ariannamethod/stb_image.h       2  16-bit convert malloc + 8-bit PNG row memcpy

stb_image.h:1820 (stbi__convert_format16) mirrors the guard its 8-bit sibling already
gets from stbi__malloc_mad3; both stb sites are hardened here as part of the vendored
copy. Same class + files as canonical notorch (#24/#25).

Proof (neo): cc -O2 -Wall -Wextra -c ariannamethod/notorch.c (Accelerate) exit 0, no new
warnings (baseline 3 == after 3); a TU including notorch_vision.h (which compiles the stb
implementation) compiles clean. Numerically inert for in-range sizes. Based on current
origin/main.

Quote: "Round numbers are always false." — Samuel Johnson (Boswell's Life of Johnson)
Method: the Method keeps every product honest in the wide word, so no rounded int stands in for a size it cannot hold.

Co-authored by Claude Opus 4.8 (Arianna Method, neo) <theariannamethod@gmail.com>, Coordinated with maintainer @iamolegataeff

+++ ariannamethod/notorch_vision.h
@@ -43,7 +43,7 @@ static nt_image* nt_image_load(const char* path, int desired_channels) {
     img->width = w;
     img->height = h;
     img->channels = c;
-    img->data = (float*)malloc(c * h * w * sizeof(float));
+    img->data = (float*)malloc((size_t)c * h * w * sizeof(float));
 
     // Convert uint8 HWC → float CHW, normalized to [0, 1]
     for (int ch = 0; ch < c; ch++)
@@ -66,7 +66,7 @@ static nt_image* nt_image_load_mem(const unsigned char* buf, int len, int desire
     img->width = w;
     img->height = h;
     img->channels = c;
-    img->data = (float*)malloc(c * h * w * sizeof(float));
+    img->data = (float*)malloc((size_t)c * h * w * sizeof(float));
 
     for (int ch = 0; ch < c; ch++)
         for (int y = 0; y < h; y++)
@@ -94,7 +94,7 @@ static nt_image* nt_image_resize(const nt_image* src, int tw, int th) {
     dst->width = tw;
     dst->height = th;
     dst->channels = src->channels;
-    dst->data = (float*)malloc(src->channels * th * tw * sizeof(float));
+    dst->data = (float*)malloc((size_t)src->channels * th * tw * sizeof(float));
 
     float sx = (float)src->width / tw;
     float sy = (float)src->height / th;
@@ -138,7 +138,7 @@ static nt_image* nt_image_center_crop(const nt_image* src, int tw, int th) {
     dst->width = tw;
     dst->height = th;
     dst->channels = src->channels;
-    dst->data = (float*)malloc(src->channels * th * tw * sizeof(float));
+    dst->data = (float*)malloc((size_t)src->channels * th * tw * sizeof(float));
 
     for (int ch = 0; ch < src->channels; ch++) {
         const float* sc = src->data + ch * src->height * src->width;
